+++ js/integer.js
@@ -1,0 +1,10 @@
+const getInteger = (min, max) => {
+  if (max <= min || max <= 0 || min < 0) {
+    return 'Некорректный ввод'
+  }
+  min = Math.ceil(min);
+  max = Math.floor(max);
+  return Math.floor(Math.random() * (max - min + 1)) + min;
+};
+
+export {getInteger};
